Updated outline.py to hadle not existing files
Taken from a newer version of the latex suite.

The problem is with subimports. As they change the path for further
inputs/subincludes, the script does not find them. Before it produced
an error, now it justs returns an empty content for them. The problem is
not solved, but at least it does not break if we just import a figure.